Update pages and components
