+++ client/src/components/Icon/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 
 // Existing components and icons
 import { Icon as ChakraIcon } from "@chakra-ui/react";
-import { BsArrowClockwise, BsArrowRight, BsBarChart, BsBox, BsCalendarWeek, BsCheck, BsChevronDoubleLeft, BsChevronDoubleRight, BsChevronDown, BsChevronLeft, BsChevronRight, BsChevronUp, BsExclamationOctagon, BsFillExclamationTriangleFill, BsGraphUp, BsGrid, BsInfoCircle, BsLink45Deg, BsList, BsPencilSquare, BsPlus, BsQuestionOctagon, BsSearch, BsTag, BsTextareaT, BsTrash, BsX } from "react-icons/bs";
+import { BsArrowClockwise, BsArrowRight, BsBarChart, BsBox, BsCalendarWeek, BsCheckLg, BsChevronDoubleLeft, BsChevronDoubleRight, BsChevronDown, BsChevronLeft, BsChevronRight, BsChevronUp, BsExclamationOctagon, BsFillExclamationTriangleFill, BsGraphUp, BsGrid, BsInfoCircle, BsLink45Deg, BsList, BsPencilSquare, BsPlus, BsQuestionOctagon, BsSearch, BsTag, BsTextareaT, BsTrash, BsXLg } from "react-icons/bs";
 
 // Custom types
 import { IconType } from "react-icons";
@@ -23,13 +23,13 @@ const SYSTEM_ICONS: {[k: string]: IconType} = {
   "attribute": BsTag,
 
   // Signal and action icons
-  "check": BsCheck,
+  "check": BsCheckLg,
   "info": BsInfoCircle,
   "search": BsSearch,
   "add": BsPlus,
   "edit": BsPencilSquare,
   "delete": BsTrash,
-  "close": BsX,
+  "cross": BsXLg,
   "list": BsList,
   "warning": BsFillExclamationTriangleFill,
   "exclamation": BsExclamationOctagon,
