+++ client/src/database/functions.ts
@@ -1,3 +1,4 @@
+// Utility functions and libraries
 import consola from "consola";
 import _ from "lodash";
 import axios from "axios";
